# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Denuncias en Barrios",
        "version": "1.0.0",
        "description": "API REST completa para la gestión de denuncias ciudadanas en barrios. Incluye autenticación, gestión de usuarios, administración de denuncias, panel de administración y dashboard con estadísticas.",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@denunciabarrios.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://172.208.72.155:3085",
            "description": "Servidor de desarrollo"
        }
    ],
...
    "paths": {
...
        "/admin/respuestaPredeterminada": {
            "post": {
                "summary": "Asignar respuesta predeterminada a una denuncia",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["denunciaId", "tipoRespuesta"],
                                "properties": {
                                    "denunciaId": {
                                        "type": "string",
                                        "description": "ID de la denuncia"
                                    },
                                    "tipoRespuesta": {
                                        "type": "string",
                                        "enum": ["tiempo_resolucion", "no_procede", "personalizada"],
                                        "description": "Tipo de respuesta"
                                    },
                                    "subtipo": {
                                        "type": "string",
                                        "description": "Subtipo de respuesta (rapida, media, larga, etc.)"
                                    },
                                    "mensajePersonalizado": {
                                        "type": "string",
                                        "description": "Mensaje personalizado"
                                    },
                                    "tiempoEstimadoPersonalizado": {
                                        "type": "string",
                                        "description": "Tiempo estimado personalizado"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta asignada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/StandardResponse"}
                            }
                        }
                    },
                    "400": {"description": "Error en parámetros"},
                    "404": {"description": "Denuncia no encontrada"},
                    "500": {"description": "Error del servidor"}
                }
            }
        },
        "/admin/respuestaPredeterminada/plantillas": {
            "get": {
                "summary": "Obtener plantillas de respuestas predeterminadas por categoría",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "in": "query",
                        "name": "categoria",
                        "schema": {"type": "string"},
                        "description": "Categoría para filtrar plantillas"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plantillas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/StandardResponse"}
                            }
                        }
                    },
                    "404": {"description": "No hay plantillas para esta categoría"},
                    "500": {"description": "Error del servidor"}
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT de autenticación de usuario"
            },
            "AdminAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "auth-admin",
                "description": "Token JWT de autenticación de administrador"
            }
        },
        "schemas": {
            "StandardResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "description": "Código de estado HTTP",
                        "example": 200
                    },
                    "status": {
                        "type": "string",
                        "description": "Estado de la respuesta",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "description": "Mensaje descriptivo",
                        "example": "Operación exitosa"
                    },
                    "data": {
                        "type": "object",
                        "description": "Datos de la respuesta",
                        "properties": {
                            "token": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        }
    }
}